requirements.txt
pyautogui
pillow
psutil
numpy
requests
python-dotenv
filelock
PyPDF2
